requirements.txt
email-validator==1.1.3
Flask==1.1.2
flask-restplus==0.13.0
loguru==0.6.0
python-dotenv==0.17.0
Werkzeug==0.16.1
SQLAlchemy==1.3.12
Flask-Migrate==2.5.2
flake8==4.0.1
pytest==7.0.1
pytest-mock==3.7.0